==> ML/2.Load&Split.ipynb <==
# Importing Libraries
import numpy as np
import pandas as pd
import sklearn
print('scikit-learn version\n', sklearn.__version__)
from sklearn.model_selection import train_test_split
import pickle
import dtale
import matplotlib.pyplot as plt
from sklearn.preprocessing import MinMaxScaler, KBinsDiscretizer

seed=41

# Load Dataset

df = pd.read_pickle('../Dataset_Preprocessed.pkl').drop(columns=['InChIKey', 'Chemical Formula'])

df.info()

dtale.show(df[['Kingdom', 'Superclass', 'Class', 'Subclass']].drop_duplicates(), host='localhost')

uclass = df[['Superclass', 'Class']].drop_duplicates()['Superclass'].value_counts()
pickle.dump(list(uclass[uclass==1].index), open('Dataset/SC_wunique_child.pkl', 'wb')) #Superclasses w unique child
uclass = df[['Class', 'Subclass']].drop_duplicates()['Class'].value_counts()
pickle.dump(list(uclass[uclass==1].index), open('Dataset/C_wunique_child.pkl', 'wb')) #Classes w unique child
all_categ = {'Kingdom': np.unique(df.iloc[:, 0]), 'Superclass': np.unique(df.iloc[:, 1]), 
               'Class': np.unique(df.iloc[:, 2]), 'Subclass': np.unique(df.iloc[:, 3])}
pickle.dump(all_categ, open('Dataset/all_categories.pkl', 'wb')) #all categories in the dataset

columns = list(df.columns)
print(', '.join(list(df.columns)), end='\n\n')
df.info()

# Split data to Train and Test

feature_names = list(df.columns)[4:] #all features
pickle.dump(feature_names, open('Dataset/feature_names.pkl', 'wb'))

#Splitting the data into independent and dependent variables
X = df.iloc[:, 4:].values
y = df.iloc[:, 0:4].values

#Creating the Training and Test set from data
X_train, X_test, y_train, y_test = train_test_split(X, y, stratify = y, test_size=0.33, random_state = seed)

pickle.dump(X_train, open(f'Dataset/X_train.pkl', 'wb'))
pickle.dump(X_test, open(f'Dataset/X_test.pkl', 'wb'))
pickle.dump(y_train, open(f'Dataset/y_train.pkl', 'wb'))
pickle.dump(y_test, open(f'Dataset/y_test.pkl', 'wb'))

### Scaling (final preprocessing step)

X_train = pickle.load(open(f'Dataset/X_train.pkl', 'rb'))
X_test = pickle.load(open(f'Dataset/X_test.pkl', 'rb'))
y_train = pickle.load(open(f'Dataset/y_train.pkl', 'rb'))
y_test = pickle.load(open(f'Dataset/y_test.pkl', 'rb'))

X_train_df = pd.DataFrame(X_train, columns=feature_names)

%matplotlib inline
for i in range(len(feature_names)):
    X_train_df.iloc[:, i].plot.hist(bins=300)
    plt.show()

scaler = MinMaxScaler(feature_range=(0,1), )
scaler.fit(X_train)
pickle.dump(scaler, open('Dataset/fitted_scaler.pkl', 'wb'))

X_train_scal = scaler.transform(X_train)
X_test_scal = scaler.transform(X_test)
pickle.dump(X_train_scal, open(f'Dataset/X_train_scal.pkl', 'wb'))
pickle.dump(X_test_scal, open(f'Dataset/X_test_scal.pkl', 'wb'))
